+++ main/release/releasability_check.py
@@ -1,6 +1,6 @@
 from release.utils.burgr import Burgr
 from release.utils.github import GitHub
-from release.utils.releasability import Releasability
+from release.releasability import Releasability
 from release.utils.release import releasability_checks
 from release.vars import burgrx_url, burgrx_user, burgrx_password
